performance optimizations
