+++ src/input/parse_json.rs
@@ -1,8 +1,6 @@
 use std::fmt;
 
 use num_bigint::BigInt;
-use pyo3::prelude::*;
-use pyo3::types::{PyDict, PyList};
 use serde::de::{Deserialize, DeserializeSeed, Error as SerdeError, MapAccess, SeqAccess, Visitor};
 
 use crate::data_value::DataValue;
@@ -11,52 +9,6 @@ use crate::lazy_index_map::LazyIndexMap;
 pub type JsonInput = DataValue;
 pub type JsonArray = Vec<JsonInput>;
 pub type JsonObject = LazyIndexMap<String, JsonInput>;
-
-impl ToPyObject for DataValue {
-    fn to_object(&self, py: Python<'_>) -> PyObject {
-        match self {
-            Self::Null => py.None(),
-            Self::Bool(b) => b.into_py(py),
-            Self::Int(i) => i.into_py(py),
-            Self::BigInt(b) => b.to_object(py),
-            Self::Uint(i) => i.into_py(py),
-            Self::Float(f) => f.into_py(py),
-            Self::String(s) => s.into_py(py),
-            Self::Array(v) => PyList::new(py, v.iter().map(|v| v.to_object(py))).into_py(py),
-            Self::Object(o) => {
-                let dict = PyDict::new(py);
-                for (k, v) in o.iter() {
-                    dict.set_item(k, v.to_object(py)).unwrap();
-                }
-                dict.into_py(py)
-            }
-            Self::Py(o) => o.clone_ref(py),
-        }
-    }
-}
-
-impl IntoPy<PyObject> for DataValue {
-    fn into_py(self, py: Python<'_>) -> PyObject {
-        match self {
-            Self::Null => py.None(),
-            Self::Bool(b) => b.into_py(py),
-            Self::Int(i) => i.into_py(py),
-            Self::BigInt(b) => b.to_object(py),
-            Self::Uint(i) => i.into_py(py),
-            Self::Float(f) => f.into_py(py),
-            Self::String(s) => s.into_py(py),
-            Self::Array(v) => PyList::new(py, v.into_iter().map(|v| v.into_py(py))).into_py(py),
-            Self::Object(o) => {
-                let dict = PyDict::new(py);
-                for (k, v) in o {
-                    dict.set_item(k, v.into_py(py)).unwrap();
-                }
-                dict.into_py(py)
-            }
-            Self::Py(o) => o.clone_ref(py),
-        }
-    }
-}
 
 impl<'de> Deserialize<'de> for DataValue {
     fn deserialize<D>(deserializer: D) -> Result<DataValue, D::Error>
@@ -184,9 +136,9 @@ impl<'de> Deserialize<'de> for DataValue {
                         while let Some((key, value)) = visitor.next_entry()? {
                             values.insert(key, value);
                         }
-                        Ok(DataValue::Object(values))
+                        Ok(DataValue::Object(Box::new(values)))
                     }
-                    None => Ok(DataValue::Object(LazyIndexMap::new())),
+                    None => Ok(DataValue::Object(Box::new(LazyIndexMap::new()))),
                 }
             }
         }
